Explore OCI cli basics
Simple script I used to do a refresher on bash, jq and get a bit hands on with OCI apis/cli. Script works but doesn't answer my question of what is the WW OCI footprint? This is because, as I learned, OCI has a region subscription model. Means that my free tier account can only query a single region and there doesn't seem to be any other programmatic way of doing this. :/